add rollbar
https://app.asana.com/0/8055896560428/22055143121719

@razorman8669 note that you’ll want to add NODE_ENV=development to your
.env to report the correct environment locally.

now, i’m aware that the development environment for Play is “dev”, not
“development”, but the full spelling is the convention for Node:
http://stackoverflow.com/a/16979503/56817

Maybe we can rename the one in Play? I know there are a few places that
check for “dev” instead of “development” but maybe refactoring is not a
big deal. It’s a little confusing to see both “Dev” and “Development”
in Rollbar.