support editing community name and intention
note that the response is modified to return an empty object. this is
for the sake of keeping the message short, since a 200 already
indicates to the client that the value it submitted was accepted, so it
doesn’t need the server to repeat the same value back.

note that this also precludes the use of Angular’s $resource.$save,
because that empties the client-side instance and fills it with only
what the response returns.