backend support for linkedin API use
this commit doesn’t just contain standard additions to the node API. it
also supports the UI flow for authorizing LinkedIn, using a popup
window.

The UI calls /noo/linkedin/authorize in a popup window, which redirects
to linkedin.com for the user to authorize; then it redirects to
/noo/linkedin/provide, which gets an OAuth2 access token and uses it to
look up the user’s public profile url.

That url is then used in an EJS template which just contains a snippet
of javascript to be executed in the popup window. `window.opener` is
used to pass the url back into the parent window’s code, where it is
stored, to be uploaded if the user clicks Save, or thrown away if they
click Cancel.