fix Membership.activeCommunityIds
to return ints

…which uncovered a strange issue in which i was using that method as if
it were synchronous, and it was somehow working in a knex query builder
whereIn call? i guess there is logic in knex to automatically resolve
promises passed as query clause arguments… but only if they’re query
objects, not promises in general, because once i added a .then to
activeCommunityIds to parse strings as ints, the magic stopped working.