rename AWS_S3_CONTENT_URL to ASSET_HOST_URL
to reflect its true purpose, which is to specify where to fetch static
assets from. more responsibility is being given to
StaticPageController, because now that Scala’s out of the picture, we
can simplify the dev environment: instead of having hylo-frontend get
all the requests first and then broker some to hylo-node in
development, we can restructure things so that it’s like it is in
production: requests go to node first and some of them get proxied over
to frontend. the only difference in development is that we still
request Angular templates individually instead of bundling them into
the JS, so we extend StaticPageController here to support that.