add alternate means of talking to popup opener during login/signup
with hylo-redux (at least in development) the auth popup ends up with a
different domain than the frontend, which causes the opener.popupDone
function call to be blocked due to a cross-origin security policy. so
we use opener.postMessage instead, and pass in the parent’s domain
through a request parameter stored in the session.