upgrade knex and bookshelf
doing this so i can use some new API methods — however, one of the
changes is that having an undefined value in a SQL query binding is now
an error, so there are lots of changes in this commit to avoid doing
that.

one particular annoying spot is that if you’ve just saved a Bookshelf
model instance, it won’t yet have populated its attributes with null
values for null foreign keys, e.g. an Invitation that does not have a
Tag. If you then try to do `invitation.load(‘tag’)` you will get an
error about an undefined binding. So instead you can do
`invitation.refresh({withRelated: ‘tag’})`.