don't assume CommunityTag exists when a user is following a tag
this attempts to cover the edge case we’re about to create, in which a
tag doesn’t “officially” exist in a community, e.g. because it was
edited into a post that was in multiple communities, by someone who was
a member of only some of them. fortunately our denormalized schema,
where TagFollow contains community_id and tag_id rather than
community_tag_id, makes this relatively straightforward.